Bug Fix: fix incorrect usage of "sizeof(blockhdr_t)".
The size of the block header differs depending on the version of the image.
sizeof(blockhdr_t) gives the size of the current version of the header when
the binary was built (V4 currently). This only mattered in two places:

In imagezip, it would have prevented us from creating a backward compatible
V1 image. Probably nobody has done this in 10 years.

In imageunzip, it could prevent relocations from being processed. This only
affected recent builds (V4 format) when attempting to unzip an older
single-partition FreeBSD image (FBSD5 or before). This is unlikely to occur
outside of Utah and only very rarely here.