Client half of the fetch-tarballs-via-the-web change.
For every tarball and rpm, tmcd will now pass a SERVER= string as well
telling the client where the file should be downloaded from (if using
the web rather than NFS). Right now this value is the same for all
tarballs and rpms, and is hardwired in tmcd as either "www" (if
SPEWFROMOPS=0) or "users" (if 1). Note: BUMPED THE TMCC VERSION NUMBER
for this.

Made a common routine for doing an error-check-and-retry copy of a file
across "racy" NFS. This is used by install-{tarfile,rpm} and rc.topomap.