Command line option to set the socket buffer size, fix some trace logic.
The "-k" option (hey, it was the best available letter!) takes a size
in MB with which the frisbee client and server will first try to size their
UDP socket buffers. They may still wind up with a smaller size, depending
on what the OS supports.

Fixed the burst logging logic so that we correctly trace overrun conditions.