Make socket buffer size more configurable.
It appeared to be before, but wasn't really. The -k option for both client
and server will set the max socketbuf size in KB (NOTE: THIS USED TO BE MB!)
The actual socketbuf size will then be the min of that and what the system
supports.

The client stats now include the sockbuf size of the run.