Get rid of the last non-used of a condition variable.
If the decompressor thread had nothing to do, it used to just sleep for
a fixed period (1ms) and check again. If CONDVARS_WORK is set, we now just
sleep on a condvar and wait for the network thread to wake us up when it
finishes filling out a chunk.