Add stat to keep track of "partial chunk drops".
These are drops of the so-called dubious chunks. Dubious chunks are those
which we partially filled, but then we started receiving pieces of another
chunk before completing the first. We mark that first chunk as dubious under
the assumption that the remainder of that chunk got dropped (by us or on
the wire) and we won't be seeing the remaining blocks for some time.