Add frisbee master server mechanisms for turning on dynamic rate tuning.
For the Emulab configuration, we add the new site variable
"images/frisbee/maxrate_dyn" which should be set non-zero to enable
dynamic adjustment. If maxrate_dyn is enabled, then the maxrate_{std,usr}
values are used as both the initial and maximum values for the BW of any
instance. Really, if maxrate_dyn is on, then both of those should be set
to the same value so that all servers are operating the same and the value
should be just above the link BW.

For the "null" configuration (aka, the subboss configuration),
this is set by adding command line options:
    -O dynamicbw=1,bandwidth=1100000000
which would enable it and start/cap the BW at 1.1Gb/sec.