Add "-T <sec>" server-timeout option to frisbee client.
If we don't get any messages from our server after <sec> seconds,
we exit. Odd that this didn't already exist, huh?

I re-purposed the -T option which used to be used to specify a prefix for
a trace file. Since I am the only one that ever used that option, I don't
feel bad about changing it. Traceprefix is now specified with -U.