Add support for auto-detecting when to create a delta image.
Uses the new-ish imagezip -P option. But first we had to get that option
to imagezip. So the key-value interface between server and client now has
an IZOPTS key where you can pass options to imagezip (This is now used to
pass the -N that we have always specified as well).

Also changed imagezip to print out a less ambiguous message when the option
is in use, if will clearly say that it is full or delta. This matters because
we have to parse the imagezip output on the server to figure out what kind
of image it is!