If there are no valid data to save, exit with an error.
E.g., telling it to save an unused (type 0) partition. Previously, we
would create a single valid range with start at the beginning of disk
(or partition) and a size of zero. However, size zero is taken by the
imagezip compressor to mean "til EOF" so we would wind up creating a
really large image instead of doing nothing.

It is debatable whether this is really an error. Perhaps we should just
create a valid image file with one chunk containing no valid ranges, but
it seems likely that they probably made a mistake and we should call it
out immediately.