Attempt to prevent progmode capture from hanging on program death.
I was attempting to read back any last words the program might have
uttered, but if it said nothing, we would hang. I would not have
expected this behavior from a pipe (actually, socketpair) when the
other end has gone away! But, make it non blocking before we read
to be safe.