Double the timeout period for waiting for the master server to reply.
This is a "temporary" hack to get around the single-threaded nature of
the master server. That in itself is not the killer, it is that it
sleeps for 2 seconds after it spawns a child process (frisbeed, frisbee
or uploader). So if three images are requested simultaneously by a handful
of clients, then it will take 6 seconds to get the frisbeed processes
started, exceeding the 5 second timeout in the clients. So other clients
in the queue after the three clients that are starting images will timeout
before the master server handles them.