Make sure partition code doesn't add skips outside the desired range.
The partition code (MBR, GPT) was adding skips to the list for ranges
that were outside of any partition. That is fine if we are making a whole
disk image, but if were are making a single-partition image it would add
sectors that were outside of the expected range triggering an assert.

This commit affects a lot of files because I modified the API to the
partition code, doing what I should have done originally. This actually
cleaned up some code cooties.