Hopefully fix frisbee uploader "no such file or directory" errors.
When using AMD, the uploader path wound up "realpath"ed in the form
of /.amd_mnt/ops/... which is the location at which AMD does the NFS
mount when triggered. However, if the mount times out, that path is
no longer valid.

So for the AMD case, we have to strip the AMD prefix from the path.
This ensures that subsequent stats and other accesses of the path go
through the AMD mountpoint (e.g., "/proj/foo") and not the NFS mountpoint
(e.g., "/.amd_mnt/ops/proj/foo"), and thus trigger AMD to do the NFS mount.