Add the ability to "report" for proxied nodes.
Not even 48 hours old and I have already had to change it...
Forgot about vnode hosts that proxy on behalf of their vnodes.
For those, we want to be reporting the actual client identity and
not the physical host's. So add a "who" field to the report message
to explicitly identify who requested the data. Previously we were
just using the IP of the sender of the report message to identify
the client.

Note that this is NOT backward compatible with yesterday's version.
Since I know everywhere that version was running, I could just
eliminate them and pretend that version never existed. Dead men
tell no tales.