Redo the frisbee heartbeat code again.
My "shortcut" to enable a heartbeat via a client-side command line
proved to be untenable. There are just too many places where we fire
off the client and getting the right heartbeat interval value to all
those places would have been...challenging.

So back to the original plan of having a server-side command line
option and letting the server tell the client when/what to report.
This limits the changes to just the frisbee master server, in particular
I now just have to get the value to master server instances running
on the subbosses (not done yet, just hardwiring a value for now).

All this said, I still had to modify the various places we invoke
the frisbee client to add an option to enable the heartbeat, but at
least I didn't need to know a specific value.