Limit setting of MBR partition sizes to 2TB.
In the 64-bit MFS, the disk size would be returned as a 64-bit value
which we would try to shoehorn into a 32-bit MBR field. The resulting
setup of partition 4 on a > 2TB disk would not be pretty.

NOTE: this limits us to using only a little over 2TB (~18GB in partitions
1-3 and 2TB in partition 4) of large disks right now.