Add Docker serial console support.
We do this similarly to Xen.  There's a new script (container2pty.py)
that attaches to the Docker container, via the docker daemon, and
exports its stdio as a pty.  Then we run capture on a symlink to that
pty.  New options to capture tell it to keep retrying to open the pty
maxretries times (we invoke with infinitely many retries); and to not
prepend /dev to the device string.