Remove any imagezip.stat file before we execute the command.
The "basic shell sleezy trick" we do only creates that file if the
imagezip fails. However, if it succeeded but there happened to be
an imagezip.stat file left-over from a previous run, we would think
we failed. This would only be a problem when taking images from
somewhere other than in an MFS (e.g., taking an image-backed data
set image from the node itself).