Initial CentOS 8 clientside support.
We continue to use the old, deprecated network-scripts/ifcfg/ifup
control net search method, since our only alternative is
NetworkManager (no systemd-networkd).  dhclient-script no longer allows
us to set "forever" for the dhcp lease time because it does math on the
value; so now we use ~UINT_MAX.  This is annoying but fine, of course.

CentOS no longer provides any ntpd, so that is absent from this commit.
Adding support for chrony is a bit more complicated, but that is what we
will do.