"15 minute hack" to propagate server-side uploader errors.
...and three days later, I am done. The server upload daemon will now
exit with specialized exit codes to indicate different errors. This
exit status is picked up by the master server and recorded in the DB
for the higher-level imaging script to pick up.

Previously, that script would only see client-side errors related to
imagezip. It would not know if the server detected that the filesystem
was full, or the image was too big, or the upload exceeded the maximum
time limit. Furthermore, the client-side would just see those errors
as a server disconnect which is what happens on a successful image
upload as well.