Fixups for occasional image backed dataset capture failures.
The so-called "storage map" in the blockstore code was losing information
due to the way it was being constructed. This required a hack to compensate,
which in turn had some failure modes. And the whole thing would fail if you
tried to capture a FreeBSD blockstore (I wonder why no one ever noticed?)

And, because I am so good, I even fixed a bug that was not there when I
started!