Create action to auto build frisbee code
this runs the existing makefiles, which have
nonstandard names due to emulab's build system

basic tests: currently a noop
TODO there is a stub to run basic tests, but
this needs to be figured out for frisbe.

if the test segment passes,
it then uploads a tar of the
binaries to the gh action's workflow artifact

the user can download this artifact, or the "same"
workflow can reference it.

a TODO is to collect the "version" and "feature"
output, and annotate the item with it

imagezip has a useful 'version' flag, i
but frisbee does not yet.